mysql commands: explicit path to .my.cnf
This avois relying on the HOME envvar. Since puppet 3.0, puppet's
"execute_posix" function clears the environment, which makes these
mysql commands fail. And relying on an user related environment
variable in puppet is a bad thing anyway.

For now, the path to .my.cnf is hardcoded as it already is so in
mysql::server.